refactor(artifact-provenance): extract persisted execution proof
