+++ src/main/artifacts/provenance-repository.ts
@@ -18,7 +18,6 @@ import type {
   GetArtifactLineageRequest,
   GetArtifactVersionProvenanceRequest,
   PersistedArtifactExecutionSnapshot,
-  ProvenanceNotebookOutput,
   ProvenanceExecutionInputFile,
   ReplayArtifactVersionRequest
 } from '../../shared/artifact-provenance'
@@ -44,6 +43,10 @@ import { NotebookRunRepository } from '../notebook/repository'
 import type { NotebookRunInputFile } from '../../shared/notebook'
 import { canonicalJson, sha256, type CanonicalJson } from './provenance-canonical'
 import { ArtifactProvenanceProducerCapture } from './provenance-producer-capture'
+import {
+  parseArtifactExecutionSnapshot,
+  validateArtifactExecutionSnapshot
+} from './provenance-execution-evidence'
 import { toCheck, toReview } from '../reviewer/repository'
 import { selectReviewChainForArtifactVersion } from '../reviewer/artifact-version-review'
 import { flagStaleReviews } from '../reviewer/stale-reviews'
@@ -576,152 +579,6 @@ const stringValue = (value: unknown): string | undefined =>
 
 const numberValue = (value: unknown): number | undefined =>
   typeof value === 'number' && Number.isFinite(value) ? value : undefined
-
-const normalizeStoredProvenanceOutput = (value: unknown): ProvenanceNotebookOutput[] => {
-  const output = recordValue(value)
-  if (output?.type === 'text' && typeof output.text === 'string') {
-    return [
-      {
-        type: 'text',
-        text: output.text,
-        ...(output.truncated === true ? { truncated: true } : {})
-      }
-    ]
-  }
-  if (output?.type === 'error') {
-    const name = stringValue(output.name)
-    const message = stringValue(output.message) ?? name ?? 'Notebook execution failed.'
-    const traceback = Array.isArray(output.traceback)
-      ? output.traceback.filter((line): line is string => typeof line === 'string')
-      : typeof output.traceback === 'string' && output.traceback
-        ? output.traceback.split('\n')
-        : undefined
-    return [
-      { type: 'error', ...(name ? { name } : {}), message, ...(traceback ? { traceback } : {}) }
-    ]
-  }
-  if (output?.type === 'table') {
-    const columns = Array.isArray(output.columns)
-      ? output.columns.filter((column): column is string => typeof column === 'string')
-      : []
-    const previewRows = Array.isArray(output.previewRows)
-      ? output.previewRows.filter((row): row is unknown[] => Array.isArray(row))
-      : []
-    const rowCount = numberValue(output.rowCount)
-    return columns.length > 0 && rowCount !== undefined
-      ? [{ type: 'table', columns, rowCount, previewRows }]
-      : []
-  }
-  if (output?.type === 'omitted-media') {
-    const mimeTypes =
-      typeof output.mimeType === 'string'
-        ? [output.mimeType]
-        : Array.isArray(output.mime_types)
-          ? output.mime_types.filter((mimeType): mimeType is string => typeof mimeType === 'string')
-          : []
-    return mimeTypes.map((mimeType) => ({
-      type: 'omitted-media',
-      mimeType,
-      ...(typeof output.byteLength === 'number' ? { byteLength: output.byteLength } : {})
-    }))
-  }
-  return []
-}
-
-const parseArtifactExecutionSnapshot = (value: string): PersistedArtifactExecutionSnapshot => {
-  const parsed = JSON.parse(value) as unknown
-  const snapshot = recordValue(parsed)
-  const truncation = recordValue(snapshot?.truncation)
-  if (
-    snapshot?.schemaVersion !== 2 ||
-    typeof snapshot.rootFrameId !== 'string' ||
-    typeof snapshot.agentFrameId !== 'string' ||
-    typeof snapshot.messageBranchId !== 'string' ||
-    typeof snapshot.terminalPromptMessageId !== 'string' ||
-    typeof snapshot.producerRunId !== 'string' ||
-    typeof snapshot.producerRunIndex !== 'number' ||
-    !Number.isSafeInteger(snapshot.producerRunIndex) ||
-    typeof snapshot.createdAt !== 'string' ||
-    !Array.isArray(snapshot.inputFiles) ||
-    !Array.isArray(snapshot.runs) ||
-    (snapshot.truncation !== undefined &&
-      (!truncation ||
-        truncation.reason !== 'payload-limit' ||
-        !Number.isSafeInteger(truncation.omittedLeadingRunCount) ||
-        Number(truncation.omittedLeadingRunCount) < 0 ||
-        !Number.isSafeInteger(truncation.omittedOutputCount) ||
-        Number(truncation.omittedOutputCount) < 0 ||
-        !Number.isSafeInteger(truncation.omittedInputCount) ||
-        Number(truncation.omittedInputCount) < 0)) ||
-    snapshot.runs.some((run) => {
-      const candidate = recordValue(run)
-      return (
-        !candidate ||
-        typeof candidate.runId !== 'string' ||
-        typeof candidate.runIndex !== 'number' ||
-        !Number.isSafeInteger(candidate.runIndex) ||
-        typeof candidate.kernelKind !== 'string' ||
-        typeof candidate.script !== 'string' ||
-        typeof candidate.status !== 'string' ||
-        typeof candidate.startedAt !== 'string' ||
-        !Array.isArray(candidate.outputs) ||
-        !Array.isArray(candidate.inputFileVersionKeys)
-      )
-    })
-  ) {
-    throw new Error('Artifact Version execution snapshot schema is invalid.')
-  }
-  const normalized = parsed as PersistedArtifactExecutionSnapshot
-  return {
-    ...normalized,
-    runs: normalized.runs.map((run) => ({
-      ...run,
-      outputs: (run.outputs as unknown[]).flatMap(normalizeStoredProvenanceOutput)
-    }))
-  }
-}
-
-const validateArtifactExecutionSnapshot = (
-  snapshot: PersistedArtifactExecutionSnapshot,
-  expected: {
-    rootFrameId: string
-    agentFrameId: string
-    messageBranchId: string
-    promptMessageId: string
-    producerRunId: string | null
-    producerRunIndex: number | null
-    executionSnapshotChecksum: string
-    evidence: ArtifactVersionEvidence
-  }
-): void => {
-  const producer = expected.evidence.producer
-  const runIndexes = snapshot.runs.map((run) => run.runIndex)
-  const indexesAreStrictlyIncreasing = runIndexes.every(
-    (runIndex, index) => index === 0 || runIndex > runIndexes[index - 1]!
-  )
-  const terminalRun = snapshot.runs.at(-1)
-  if (
-    snapshot.rootFrameId !== expected.rootFrameId ||
-    snapshot.agentFrameId !== expected.agentFrameId ||
-    snapshot.messageBranchId !== expected.messageBranchId ||
-    snapshot.terminalPromptMessageId !== expected.promptMessageId ||
-    expected.producerRunId === null ||
-    expected.producerRunIndex === null ||
-    snapshot.producerRunId !== expected.producerRunId ||
-    snapshot.producerRunIndex !== expected.producerRunIndex ||
-    expected.evidence.execution_snapshot_checksum !== expected.executionSnapshotChecksum ||
-    producer.state !== 'available' ||
-    producer.producer_run_id !== expected.producerRunId ||
-    producer.run_index !== expected.producerRunIndex ||
-    snapshot.runs.length === 0 ||
-    !indexesAreStrictlyIncreasing ||
-    runIndexes.some((runIndex) => runIndex > expected.producerRunIndex!) ||
-    terminalRun?.runId !== expected.producerRunId ||
-    terminalRun.runIndex !== expected.producerRunIndex
-  ) {
-    throw new Error('Artifact Version execution snapshot metadata mismatch.')
-  }
-}
 
 type PersistedExecutionInputRow = {
   ordinal: number
